+++ priv/components/checkbox_card.exs
@@ -1,6 +1,6 @@
 [
-  radio_card: [
-    name: "radio_card",
+  checkbox_card: [
+    name: "checkbox_card",
     args: [
       variant: ["default", "outline", "shadow", "bordered", "base"],
       color: [
@@ -22,8 +22,8 @@
       space: ["extra_small", "small", "medium", "large", "extra_large", "none"],
       padding: ["extra_small", "small", "medium", "large", "extra_large"],
       rounded: ["extra_small", "small", "medium", "large", "extra_large"],
-      only: ["radio_card"],
-      helpers: [radio_card_check: 3],
+      only: ["checkbox_card"],
+      helpers: [checkbox_card_check: 3],
       module: ""
     ],
     optional: [],